move check_platform_helper_version_needs_update tests to domain
