+++ dbt_platform_helper/domain/platform_helper_version.py
@@ -1,0 +1,2 @@
+class PlatformHelperVersion:
+    pass
